wip: routing messages through hierarchy
